Update patch set 1
Patch Set 1:

Sony has moved on to using a more comprehensive solution called "system_monitor" that has a wider ambit than thermanager. (That one is not open source)

I would actually stick my neck out in saying that there wont be any contributions in the near future to this by Sony, because this is something they have completed to a level where it can take over all functions of qcom default thermald and have released it out, or sort of presented it to the community and it is more or less for the community to take this forward.

And Sony engineers have shown the goodwill to actually come and contribute on CyanogenMod/android_hardware_sony_DASH, so it wont make a lot of sense to track upstream and more sensible to maintain our own.

Whether it is to be kept here, or to be kept at hardware/sony/thermanager or at external/sony/thermanager is a call I completely let the senior contributors of CM take :)

Patch-set: 1